+++ polynominal_interpolation.h
@@ -15,14 +15,14 @@ int max(int a, int b);
 typedef struct
 {
     int size;
-    float *p;
+    double *p;
 } arr;
 
 arr *init(int n);
 arr *resize(arr *a, int new_size);
-void insert(arr *a, int pos, float val);
+void insert(arr *a, int pos, double val);
 arr *add(arr *a, arr *b);
-arr *mult(arr *a, float mul);
+arr *mult(arr *a, double mul);
 void printa(arr *a);
 arr *arr_without_el(arr *a, int ex_pos);
 arr *reverse(arr *a);
@@ -34,7 +34,7 @@ arr *reverse(arr *a);
 int has_comb(int *arr, int n, int k);
 int mult_by_index(arr *a, int *coords, int n);
 int sum_of_mult_of_n_combinations(arr *a, int n);
-int compose_denominator(arr *a, int pos);
+double compose_denominator(arr *a, int pos);
 arr *compose_interpolation_polynomial(arr *xes, arr *ys);
 
 #endif
